RxSwift를 활용한 실시간 지하철 도착정보 데이터 구성 및 UI 구성

+++ Four_Commute/Main/View/MainTableViewCell.swift
@@ -49,7 +49,7 @@ class MainTableViewCell : UITableViewCell{
         $0.axis = .vertical
     }
     
-    func cellSet(margin : Int){
+    func cellSet(margin : Double){
         self.selectionStyle = .none
         
         self.contentView.addSubview(self.mainBG)
